Update how we patch manpages
The manpages are now built with asciidoc, also do that at build time so
we can properly patch them.  Therefore build-depend on asciidoc, and
update 03_tor_manpage_in_section_8 to patch the .txt files now.  Remove
tor.1.in torify.1.in tor-gencert.1.in tor-resolve.1.in in the doc
directory during clean.