Downgrade torsocks/tsocks dependency to a recommends
Downgrade torsocks/tsocks dependency to a recommends.  That tool is not
needed if you only run a relay, or if you access Tor only using polipo
or privoxy.  The torify(1) wrapper that makes use of torsocks or tsocks
already handles their absense and emmits a proper message telling the
user what they are missing (closes: #595898).